Use original URL to determine source file name
Certain forges allow you to snag a tarball from them using a known URL.
In particular one of them will cause a redirect to an ultimate URL that
does not end in the expected filename that is present in the SourceN
line, which can cause issues of the resulting SRPM from being generated
with the actual SourceN file being present, or being an incorrect file
(ultimately leading to a failed build).

This switches real_url to url to avoid the situation where the URL for SourceN
that's being downloaded doesn't match the resulting real_url.

The other possible source of "truth" here would be to parse the headers
from the download itself, and use the content-disposition header's
filename value.  This could ALSO be wrong depending on if it's set
correctly or not.

Erring on the simpler, and more controllable to the .spec file writer of
using the base url.

An option to clarify this behavior may be worth investigating to add
some sort of explicit ultimate filename like SourceFilenameN to allow
for an explicit rename in this instance.

Signed-off-by: John 'Warthog9' Hawley (VMware) <warthog9@eaglescrag.net>